+++ scripts/browser.js
@@ -1,3 +1,5 @@
-import browser from '../src/browser'
+import { browser } from '../src/browser'
 
-browser()
+const script = document.currentScript
+
+browser(script ? script.getRootNode() : document)